[ADD/#199] 컴포즈 extension 추가
