+++ core/common/src/main/java/kr/genti/common/xml/base/BaseFragment.kt
