+++ core/common/src/main/java/kr/genti/common/xml/extension/FragmentExt.kt
@@ -1,4 +1,4 @@
-package kr.genti.common.extension
+package kr.genti.common.xml.extension
 
 import android.view.View
 import android.widget.Toast
