+++ core/common/src/main/java/kr/genti/common/xml/extension/UriExt.kt
@@ -1,4 +1,4 @@
-package kr.genti.common.extension
+package kr.genti.common.xml.extension
 
 import android.content.ContentResolver
 import android.net.Uri
