+++ core/common/src/main/java/kr/genti/common/xml/state/UiState.kt
